+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Revision history for `plutarch-unit`
 
-## 0.1.0.0 -- 2022-07-26
+## 1.1.0 -- 2022-09-21
+
+### Added
+
+* `Test.Unit.Precompiled` module for testing precompiled scripts.
+* `TestCompiled` provides monadic interface for precompiled script checking.
+* `Expectation` defines if a test case should succeed or not.
+* `(@&)` and `withApplied` stitches in shared arguments.
+* `testEvalCase`, `@>`, and `@!>` tests if script fails or succeeds given arguments.
+* `testEqualityCase` tests if script is eqaul to expectation or not given arguments.
+* `fromPTerm` compiles Plutarch term and tests in `TestCompiled`.
+
+## 1.0.0 -- 2022-07-26
 
 ### Added